src/averages/requirements.txt
attrs==19.3.0
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
dnspython==1.16.0
Flask==1.1.2
Flask-Cors==3.0.8
Flask-PyMongo==2.3.0
google-api-core==1.21.0
google-auth==1.18.0
google-cloud-bigquery==1.22.0
google-cloud-core==1.3.0
google-resumable-media==0.5.1
googleapis-common-protos==1.52.0
gunicorn==20.0.4
idna==2.10
importlib-metadata==1.6.0
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
more-itertools==8.3.0
numpy==1.18.5
packaging==20.4
pandas==1.0.5
pluggy==0.13.1
protobuf==3.12.2
py==1.8.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pymongo==3.10.1
pyparsing==2.4.7
pytest==5.4.3
python-dateutil==2.8.1
python-dotenv==0.13.0
pytz==2020.1
requests==2.23.0
rsa==4.6
six==1.15.0
urllib3==1.25.9
wcwidth==0.2.3
Werkzeug==1.0.1
zipp==3.1.0
